Add Vault-Transit client support.
Vault Transit is a secrets engine that provides "encryption as a
service": it allows creating (on the server) encryption keys with a
specific algorithm, and offers an HTTP API with endpoints to encrypt
or decrypt data, relieving client from the need to handle keys (which
is risky) or implement cryptosystems (which is error-prone).
This commit adds support for transit endpoints. In particulat:

- We add a transit-client class, made of a HTTP Client, a key name,
  and the vault base URI. The ideal use of this Transit-Client is as
  a representation of a transit connection, which hides all of the
  transit logic and only provides encryption/decryption functions.

- In line with the design of the Vault client, we also add a Transit
  object that provides the main functions separately.

- We make those transit-client endpoints generic on the data being
  encrypted or decrypted. Since transit take Base64 information, we
  add a Base64 wrapper class, and a codec (encode/decode) type-class.

- We add a set of models, to represent the requests and responses of
  the Vault-transit endpoints, and their Json encoding.