Avoid handling internal structure of CipherText
Although the examples in the Vault Transit API documentation show
CipherTexts following the format "vault:v1:{B64}", where B64 is a
base64 encoded string, there is no explicit statement as to whether
that is always the case. Dealing with that internal structure may
be unsound.

Thus, we avoid handling that label in the encoding/decoding, and we
simply pass the strings from `ciphertext` unaltered. This also allow
us to unify the classes CipherText and Encrypted.