Drop the Base64Codec type class.
Some code review comments were... sceptical, about having the
Base64Codec type class. This was founded on the fact that:

- If this library is just a Scala representation of Vault api, there
  is no need to do more than that.

- Other libraries like scodec already define such type-class with
  a rich set of instances, so no point in inventing a training wheel.

For this reason, in this commit we drop this type class and reduce the
API to simple Base64 strings.

To retain some type safety, to tell apart the Base64 strings that go
for context and for the plaintext, we use a couple of wrapper classes.