Transit: Batch Encrypt and Decrypt functions
Vault Transit endpoints to encrypt and decrypt data allow for a
batch-input mode, in which multiple plaintexts (accompanied or not by
a context) may be given in a single go and decrypted, using a single
round-trip for the operation.

This commit adds support in the Transit package for the batched mode
of the decrypt and encrypt operations.
Unlike the single-input variants, that either succeed or fail, a
batched request may succeed for some inputs but fail for others.
Thus, we need to expose the TransitError case class outside.

We also extend the documentation of the other functions, and we add
into the error some context for the operation that failed.