Transit: Batch encrypt and decrypt (#55)
The encrypt (resp. decrypt) methods of Vault Transit have a 
batch mode, in which several plaintexts (resp. ciphertexts)
and contexts are sent to the transit in a single request
and thus a single round-trip. 

This commit adds support for batched mode in transit operations 
to encrypt or decrypt. Unlike the single-input variants, 
that either succeed or fail, a batched request may succeed 
for some inputs but fail for others.
Thus, we need to expose the TransitError case class outside.

We also extend the documentation of the other functions, and we add
into the error some context for the operation that failed.