Ignore errors when revoking the lease. (#92)
The methods `readSecretAndRetain` and `loginAndKeep` include a call to revoke the token or secret lease, at the end of the respective streams method. This is a resource cleanup. However, for some people this may fail. Since this may not be relevant, we add an option, to ignore errors during cleanup.